Update mvneta/e6000sw for new style DSA DTB
The new style DSA DTB is documented here for:
https://www.kernel.org/doc/Documentation/devicetree/bindings/net/dsa/dsa.txt

This new format is used by the ESPRESSObin dts:
https://github.com/freebsd/freebsd/blob/5a7b66a89ee821edb7c05a71a4ba8d30c083f233/sys/gnu/dts/arm64/marvell/armada-3720-espressobin.dts

Which means there is no longer a top-level "marvell,dsa" OFW node, and
instead we need to look up the switch directly, and follow the xref if
necessary.

This fixes the mvneta_if code that looks to see if it has a switch
attached, currently it only looks for a single switch type, I am not
sure if there is a better way, but we likely need to walk the whole ofw
tree looking for xrefs to ourselves (assuming there is no function out
there to do that work for us?)

Second I've fixed the e6000sw to use the same method for finding its
switch node, with the change we also get the ability to just look for a
child named "ports" and then iterate over the ports themselves
individually. I've updated the code that parses the port information as
well to use xref to look up the referenced ethernet port to be able to
configure the fixed-link (although someone with more experience with
device tree bindings/DSA please chime in if that is incorrect).

Last but not least, I've changed the multi_chip selection scheme to not
check if the sw_addr is a multiple of 2. This was causing failures to
identify the switch correctly on my ESPRESSObin v7 because it would
incorrectly go into single chip addressing mode, which meant it was
failing to attach to the device correctly.